refactor to re-use experiment dropdown
