+++ app/src/components/table/styles.ts
@@ -50,6 +50,10 @@ export const tableCSS = (theme: Theme) => css`
             background: var(--ac-global-color-primary);
           }
         }
+        // Style action menu buttons in the header
+        .ac-button[data-size="compact"][data-childless="true"] {
+          padding: var(--ac-global-dimension-size-25);
+        }
       }
     }
   }
